Removed deprecated code.

+++ lib/gym/package_command_generator.rb
@@ -67,29 +67,27 @@ module Gym
         developer_dir = `xcode-select --print-path`.strip
         patched_package_application_path = File.join("/tmp", "PackageApplication4Gym")
         # Remove any previous patched PackageApplication
-        FileUtils.rm patched_package_application_path
+        FileUtils.rm patched_package_application_path if File.exist?(patched_package_application_path)
 
-        Dir.mktmpdir do |tmpdir|
-          # Check current PackageApplication MD5
-          require 'digest'
+        # Check current PackageApplication MD5
+        require 'digest'
 
-          expected_md5 = File.read("lib/assets/package_application_patches/PackageApplication_MD5")
+        expected_md5 = File.read("lib/assets/package_application_patches/PackageApplication_MD5")
 
-          raise "Found an invalid `PackageApplication` script. This is not supported." unless expected_md5 == Digest::MD5.file("#{developer_dir}/Platforms/iPhoneOS.platform/Developer/usr/bin/PackageApplication").hexdigest
+        raise "Found an invalid `PackageApplication` script. This is not supported." unless expected_md5 == Digest::MD5.file("#{developer_dir}/Platforms/iPhoneOS.platform/Developer/usr/bin/PackageApplication").hexdigest
 
-          # Duplicate PackageApplication script to PackageApplication4Gym
-          FileUtils.copy_file("#{developer_dir}/Platforms/iPhoneOS.platform/Developer/usr/bin/PackageApplication", patched_package_application_path)
+        # Duplicate PackageApplication script to PackageApplication4Gym
+        FileUtils.copy_file("#{developer_dir}/Platforms/iPhoneOS.platform/Developer/usr/bin/PackageApplication", patched_package_application_path)
 
-          # Apply patches to PackageApplication4Gym from patches folder
-          Dir["lib/assets/package_application_patches/*.diff"].each do |patch|
-            puts "Applying Package Application patch: #{File.basename(patch)}"
-            command = ["patch #{patched_package_application_path} < #{patch}"]
-            print_command(command, "Applying Package Application patch: #{File.basename(patch)}") if $verbose
+        # Apply patches to PackageApplication4Gym from patches folder
+        Dir["lib/assets/package_application_patches/*.diff"].each do |patch|
+          puts "Applying Package Application patch: #{File.basename(patch)}"
+          command = ["patch #{patched_package_application_path} < #{patch}"]
+          print_command(command, "Applying Package Application patch: #{File.basename(patch)}") if $verbose
 
-            Gym::CommandsExecutor.execute(command: command, print_all: false, error: proc do |output|
-              ErrorHandler.handle_build_error(output)
-            end)
-          end
+          Gym::CommandsExecutor.execute(command: command, print_all: false, error: proc do |output|
+            ErrorHandler.handle_build_error(output)
+          end)
         end
 
         patched_package_application_path
